Reservar Almuerzo Listo 80%

+++ SmartFoodV2.0/SmartFood/src/smartfood/classes/user/Usuario.java
@@ -117,7 +117,7 @@ public class Usuario {
         this.activo = activo;
     }
     
-    public boolean pagar(float pago, PagoStrategy strategy) {
+    public boolean pagar(double pago, PagoStrategy strategy) {
         return strategy.pagar(pago);
     }
