+++ SmartFoodV2.0/SmartFood/src/smartfood/models/CarneInteligenteStrategy.java
@@ -32,7 +32,7 @@ public class CarneInteligenteStrategy implements PagoStrategy {
     }
 
     @Override
-    public boolean pagar(float pago) {
+    public boolean pagar(double pago) {
         if (this.verificarCredito(pago)) {
             this.saldo -= pago;
             return true;
@@ -41,7 +41,7 @@ public class CarneInteligenteStrategy implements PagoStrategy {
     }
 
     @Override
-    public boolean verificarCredito(float pago) {
+    public boolean verificarCredito(double pago) {
         return this.saldo - pago >= 0;
     }
